fix: `can't create user record: duplicated key not allowed` errors
If a user issued multiple concurrent requests to the API server this 
could occur if the user had never use the api before.
